requirements.txt
bcrypt==3.1.3
cffi==1.10.0
Django==1.11.3
django-extensions==1.8.1
pycparser==2.18
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
six==1.10.0
